+++ crates/gamai/src/prop/mod.rs
@@ -1,12 +1,14 @@
-mod child_prop;
-pub use self::child_prop::*;
 mod action_result;
 pub use self::action_result::*;
+mod action_timer;
+pub use self::action_timer::*;
+mod child_prop;
+pub use self::child_prop::*;
 mod prop;
 pub use self::prop::*;
-mod tree_bundle;
-pub use self::tree_bundle::*;
 mod prop_tree;
 pub use self::prop_tree::*;
 mod score;
 pub use self::score::*;
+mod tree_bundle;
+pub use self::tree_bundle::*;