+++ crates/sweet/src/bevy_ecs_matchers/mod.rs
@@ -1,3 +1,5 @@
+mod app_x;
+pub use self::app_x::*;
 mod sweet_into;
 pub use self::sweet_into::*;
 mod world;