+++ docs/src/gamai/changelog.md
@@ -2,6 +2,10 @@
 
 ## TODO
 - interrupts (recursive cleanup)
+- ActionTimers
+
+## main
+- private action inners
 
 ## `0.1.32` - 17/10/2023
 - trees